[gles/vulkan] moved around headers [misc] updated readme
