+++ core/CMakeLists.txt
@@ -22,15 +22,17 @@ if(VS_USE_NATVIS)
 endif()
 link_libraries(${WSI_LIB})
 
-FetchContent_Declare(vulkan GIT_REPOSITORY https://github.com/KhronosGroup/Vulkan-Headers GIT_TAG v${VK_VERSION} GIT_SHALLOW TRUE GIT_PROGRESS TRUE)
-FetchContent_GetProperties(vulkan)
-if(NOT vulkan_POPULATED)
-	message(STATUS "fetching KhronosGroup/Vulkan-Headers...")
-	FetchContent_Populate(vulkan)
-	if(VULKAN_STATIC)	
-		add_subdirectory(${vulkan_SOURCE_DIR} ${vulkan_BINARY_DIR})
+if(PE_VULKAN)
+	FetchContent_Declare(vulkan GIT_REPOSITORY https://github.com/KhronosGroup/Vulkan-Headers GIT_TAG v${VK_VERSION} GIT_SHALLOW TRUE GIT_PROGRESS TRUE)
+	FetchContent_GetProperties(vulkan)
+	if(NOT vulkan_POPULATED)
+		message(STATUS "fetching KhronosGroup/Vulkan-Headers...")
+		FetchContent_Populate(vulkan)
+		if(VK_STATIC)	
+			add_subdirectory(${vulkan_SOURCE_DIR} ${vulkan_BINARY_DIR})
+		endif()
+		message(STATUS "KhronosGroup/Vulkan-Headers fetched")
 	endif()
-	message(STATUS "KhronosGroup/Vulkan-Headers fetched")
 endif()
 
 FetchContent_Declare(gli GIT_REPOSITORY https://github.com/g-truc/gli.git GIT_TAG 0.8.2.0 GIT_SHALLOW TRUE GIT_PROGRESS TRUE)
@@ -41,34 +43,35 @@ if(NOT gli_POPULATED)
 	message(STATUS "g-truc/gli fetched")
 endif()
 
-
-if(VULKAN_STATIC)		
-	link_directories(${vulkan_SOURCE_DIR}/install)
-else()	
-	FetchContent_Declare(volk GIT_REPOSITORY https://github.com/zeux/volk.git GIT_PROGRESS TRUE)
-	FetchContent_GetProperties(volk)
-	if(NOT volk_POPULATED) 
-		message(STATUS "fetching volk...")
-		FetchContent_Populate(volk)
-		
-		message(STATUS "generating dynamic lib for vulkan")
-		set(PY_RES 0)
-		EXECUTE_PROCESS(WORKING_DIRECTORY "${volk_SOURCE_DIR}" RESULT_VARIABLE PY_RES COMMAND python3 "generate.py" ${vulkan_SOURCE_DIR}/registry/vk.xml)
-		if(NOT ${PY_RES} EQUAL 0)
-		  EXECUTE_PROCESS(WORKING_DIRECTORY "${volk_SOURCE_DIR}" RESULT_VARIABLE PY_RES COMMAND python "generate.py" ${vulkan_SOURCE_DIR}/registry/vk.xml)
-		endif()
-		message(STATUS "generated dynamic lib")
-		
-		set_source_files_properties("${volk_SOURCE_DIR}/volk.c" PROPERTIES LANGUAGE CXX )
-		
-		project(volk_proj VERSION 1.0.0 LANGUAGES CXX)
-		add_library(volk_proj STATIC ${volk_SOURCE_DIR}/volk.h ${volk_SOURCE_DIR}/volk.c)
-		target_compile_options(volk_proj PRIVATE ${PE_COMPILE_OPTIONS} ${PE_COMPILE_OPTIONS_LIB})
-		target_compile_features(volk_proj PRIVATE ${PE_COMPILER_FEATURES})
-		set_target_properties(volk_proj PROPERTIES LINKER_LANGUAGE CXX)
-		target_include_directories(volk_proj PUBLIC ${vulkan_SOURCE_DIR}/include)
-		message(STATUS "volk fetched")
-	endif()	
+if(PE_VULKAN)
+	if(VK_STATIC)		
+		#link_directories(${vulkan_SOURCE_DIR}/install)
+	else()	
+		FetchContent_Declare(volk GIT_REPOSITORY https://github.com/zeux/volk.git GIT_PROGRESS TRUE)
+		FetchContent_GetProperties(volk)
+		if(NOT volk_POPULATED) 
+			message(STATUS "fetching volk...")
+			FetchContent_Populate(volk)
+			
+			message(STATUS "generating dynamic lib for vulkan")
+			set(PY_RES 0)
+			EXECUTE_PROCESS(WORKING_DIRECTORY "${volk_SOURCE_DIR}" RESULT_VARIABLE PY_RES COMMAND python3 "generate.py" ${vulkan_SOURCE_DIR}/registry/vk.xml)
+			if(NOT ${PY_RES} EQUAL 0)
+				EXECUTE_PROCESS(WORKING_DIRECTORY "${volk_SOURCE_DIR}" RESULT_VARIABLE PY_RES COMMAND python "generate.py" ${vulkan_SOURCE_DIR}/registry/vk.xml)
+			endif()
+			message(STATUS "generated dynamic lib")
+			
+			set_source_files_properties("${volk_SOURCE_DIR}/volk.c" PROPERTIES LANGUAGE CXX )
+			
+			project(volk_proj VERSION 1.0.0 LANGUAGES CXX)
+			add_library(volk_proj STATIC ${volk_SOURCE_DIR}/volk.h ${volk_SOURCE_DIR}/volk.c)
+			target_compile_options(volk_proj PRIVATE ${PE_COMPILE_OPTIONS} ${PE_COMPILE_OPTIONS_LIB})
+			target_compile_features(volk_proj PRIVATE ${PE_COMPILER_FEATURES})
+			set_target_properties(volk_proj PROPERTIES LINKER_LANGUAGE CXX)
+			target_include_directories(volk_proj PUBLIC ${vulkan_SOURCE_DIR}/include)
+			message(STATUS "volk fetched")
+		endif()	
+	endif()
 endif()
 
 SET(GLES_FILES ${CMAKE_CURRENT_SOURCE_DIR}/../external/glad/src/glad.c ${CMAKE_CURRENT_SOURCE_DIR}/../external/wgl/src/glad_wgl.c)
@@ -81,7 +84,7 @@ else()
 	add_library(${LOCAL_PROJECT} STATIC ${FWD} ${SRC} ${INC} ${INC_GLES} ${SRC_GLES} ${INC_VULKAN} ${SRC_VULKAN} ${NATVIS} ${GLES_FILES})
 endif()	
 
-if(VULKAN_STATIC)	
+if(VK_STATIC AND PE_VULKAN)	
 	if(WIN32)
 		set(vk_lib_name "vulkan-1")
 	else()
@@ -99,7 +102,7 @@ endif()
 target_include_directories(${LOCAL_PROJECT}
 	PUBLIC 
       ${CMAKE_CURRENT_SOURCE_DIR}/inc
-		${VULKAN_INC_PATH}
+		${vulkan_SOURCE_DIR}/include
 		${volk_SOURCE_DIR}
 		${CMAKE_CURRENT_SOURCE_DIR}/../external/glad/include
 		${CMAKE_CURRENT_SOURCE_DIR}/../external/wgl/include
