+++ core/inc/data/buffer.h
@@ -2,10 +2,10 @@
 #include "psl/serialization.h"
 #include <vector>
 #include <optional>
-#include "vk/stdafx.h"
 #include "psl/memory/region.h"
 #include "psl/memory/segment.h"
 #include "gfx/types.h"
+#include "gfx/stdafx.h"
 #include "fwd/resource/resource.h"
 
 /// \brief contains all data types that can be serialized to/from disk.
