+++ core/inc/gfx/limits.h
@@ -1,7 +1,9 @@
 #pragma once
 #include "context.h"
 #include "stdafx.h"
+#ifdef PE_VULKAN
 #include "vk/context.h"
+#endif
 
 namespace core::gfx::limits
 {
@@ -11,15 +13,19 @@ namespace core::gfx::limits
 		{
 		case graphics_backend::gles:
 		{
+#ifdef PE_GLES
 			int gl_align = 4;
 			glGetIntegerv(GL_SHADER_STORAGE_BUFFER_OFFSET_ALIGNMENT, &gl_align);
 			return static_cast<uint64_t>(gl_align);
 		}
 		break;
+#endif
+#ifdef PE_VULKAN
 		case graphics_backend::vulkan:
 			return static_cast<uint64_t>(
 				context.resource().get<core::ivk::context>()->properties().limits.minStorageBufferOffsetAlignment);
 			break;
+#endif
 		}
 		return 0;
 	}
@@ -30,16 +36,19 @@ namespace core::gfx::limits
 		{
 		case graphics_backend::gles:
 		{
-
+#ifdef PE_GLES
 			int gl_align = 4;
 			glGetIntegerv(GL_UNIFORM_BUFFER_OFFSET_ALIGNMENT, &gl_align);
 			return static_cast<uint64_t>(gl_align);
 		}
 		break;
+#endif
+#ifdef PE_VULKAN
 		case graphics_backend::vulkan:
 			return static_cast<uint64_t>(
 				context.resource().get<core::ivk::context>()->properties().limits.minUniformBufferOffsetAlignment);
 			break;
+#endif
 		}
 		return 0;
 	}
