+++ core/inc/gfx/stdafx.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #ifdef PE_VULKAN
-#include "vk/stdafx.h"
+#include "vk/ivk.h"
 #endif
 #ifdef PE_GLES
 #include "glad/glad.h"
