+++ core/inc/gles/igles.h
@@ -1,0 +1,2 @@
+#pragma once
+#include "glad/glad.h"
