+++ core/inc/vk/buffer.h
@@ -3,6 +3,7 @@
 #include "resource/resource.hpp"
 #include <optional>
 #include "psl/memory/segment.h"
+#include "vk/ivk.h"
 
 
 namespace core::data
