+++ core/inc/vk/geometry.h
@@ -1,5 +1,5 @@
 ﻿#pragma once
-#include "vk/stdafx.h"
+#include "vk/ivk.h"
 #include "resource/resource.hpp"
 #include "psl/memory/range.h"
 #include "psl/memory/segment.h"
