+++ core/src/gles/geometry.cpp
@@ -7,6 +7,7 @@
 #include "gles/shader.h"
 #include "psl/array.h"
 #include "logging.h"
+#include "gles/igles.h"
 
 using namespace core::igles;
 using namespace core::resource;
