+++ core/src/vk/swapchain.cpp
@@ -2,6 +2,7 @@
 #include "os/surface.h"
 #include "vk/context.h"
 #include "vk/texture.h"
+#include "vk/conversion.h"
 #include "meta/texture.h"
 #include "logging.h"
 
@@ -381,7 +382,7 @@ void swapchain::init_depthstencil()
 	metaData->depth(1);
 	metaData->mip_levels(1);
 	metaData->image_type(gfx::image_type::planar_2D);
-	metaData->format(to_format(depthFormat));
+	metaData->format(conversion::to_format(depthFormat));
 	metaData->usage(core::gfx::image_usage::dept_stencil_attachment | core::gfx::image_usage::transfer_source);
 	metaData->aspect_mask(core::gfx::image_aspect::depth | core::gfx::image_aspect::stencil);
 	m_DepthTextureHandle = m_Cache.create_using<core::ivk::texture>(std::move(metaData), m_Context);
@@ -407,7 +408,7 @@ void swapchain::init_renderpass()
 	attachments[0].finalLayout	= vk::ImageLayout::ePresentSrcKHR;
 
 	// Depth attachment
-	attachments[1].format		  = core::gfx::to_vk(m_DepthTextureHandle->meta().format());
+	attachments[1].format		  = core::gfx::conversion::to_vk(m_DepthTextureHandle->meta().format());
 	attachments[1].samples		  = vk::SampleCountFlagBits::e1;
 	attachments[1].loadOp		  = vk::AttachmentLoadOp::eClear;
 	attachments[1].storeOp		  = vk::AttachmentStoreOp::eStore;
