+++ core/src/vk/texture.cpp
@@ -6,6 +6,7 @@
 #include "vk/buffer.h"
 #include "data/buffer.h"
 #include "vk/sampler.h"
+#include "vk/conversion.h"
 #ifdef fseek
 #define cached_fseek fseek
 #define cached_fclose fclose
@@ -27,7 +28,7 @@
 #define ftell cached_ftell
 #endif
 using namespace psl;
-using namespace core::gfx;
+using namespace core::gfx::conversion;
 using namespace core::ivk;
 using namespace core::resource;
